feat(web): Moved addPowerProducerToFactory to power.ts
